requirements.txt
Flask
pillow
requests